replace create DGE commands as classmethods
