+++ tests/test_DGEList.py
@@ -1,11 +1,14 @@
-from smart_open import smart_open
-import numpy as np
 import pytest
 
 from edgePy.DGEList import DGEList
-from edgePy.data_import.data_import import get_dataset_path
-from edgePy.data_import.data_import import create_DGEList_handle
 
+import pkgutil
+
+import numpy as np
+
+from smart_open import smart_open
+
+from edgePy.data_import.data_import import get_dataset_path
 
 TEST_DATASET = "GSE49712_HTSeq.txt.gz"
 TEST_DATASET_NPZ = "GSE49712_HTSeq.txt.npz"
@@ -15,7 +18,7 @@ TEST_GROUPS = "groups.json"
 @pytest.fixture
 def dge_list():
     with smart_open(get_dataset_path(TEST_DATASET), 'r') as handle:
-        return create_DGEList_handle(handle, get_dataset_path(TEST_GROUPS))
+        return DGEList.create_DGEList_handle(handle, get_dataset_path(TEST_GROUPS))
 
 
 def test_sample_by_group():
@@ -202,3 +205,57 @@ def test_non_implemented():
         dge_list().rpkm(None)
     with pytest.raises(NotImplementedError):
         dge_list().tpm(None)
+
+
+# Unit tests for ``edgePy.data_import.Importer``.\
+def test_init():
+    dge_list = DGEList.create_DGEList_handle(
+        data_handle=smart_open(get_dataset_path(TEST_DATASET)),
+        group_file=get_dataset_path(TEST_GROUPS),
+    )
+
+    assert dge_list.__repr__() == "DGEList(num_samples=10, num_genes=21,716)"
+
+
+# TestGroupImporter.
+def test_GroupImporter_init():
+    dge_list = DGEList.create_DGEList_handle(
+        data_handle=smart_open(get_dataset_path(TEST_DATASET)),
+        group_file=get_dataset_path(TEST_GROUPS),
+    )
+    assert 2 == len(dge_list.groups_dict)
+    assert 5 == len(dge_list.groups_dict["Group 1"])
+    assert 5 == len(dge_list.groups_dict["Group 2"])
+
+    assert dge_list.samples.shape == dge_list.groups_list.shape
+
+
+# Unit tests for packaged (optionally zipped during install) data.
+def test_get_data_stream():
+    """Tests finding packaged data with ``pkgutil.get_data()``"""
+    pkgutil.get_data("edgePy", "data/GSE49712_HTSeq.txt.gz")
+
+
+def test_create_DGEList():
+    """Tests the function that converts data into a DGE_List object"""
+    samples = ["AAA", "BBB", "CCC"]
+    genes = ["ENSG001", "ENSG002"]
+
+    data_set = {
+        "AAA": {"ENSG001": 10, "ENSG002": 20},
+        "BBB": {"ENSG001": 15, "ENSG002": 40},
+        "CCC": {"ENSG001": 20, "ENSG002": 80},
+    }
+    categories = {"AAA": "One", "BBB": "Two", "CCC": "One"}
+
+    dge_list = DGEList.create_DGEList(
+        sample_list=samples, data_set=data_set, gene_list=genes, sample_category=categories
+    )
+
+    assert np.array_equal(dge_list.samples, np.array(samples))
+    # 2 rows (genes), 3 columns(samples)
+    assert np.array_equal(dge_list.counts, np.array([[10, 15, 20], [20, 40, 80]]))
+
+    assert np.array_equal(dge_list.groups_list, np.array(["One", "Two", "One"]))
+    assert dge_list.groups_dict, {"One:"}
+    assert np.array_equal(dge_list.genes, np.array(genes))
